Conflicts:
	.gitignore
	android/cli/commands/_build.js
	android/modules/accelerometer/project.properties
	android/modules/analytics/project.properties
	android/modules/android/project.properties
	android/modules/android/src/java/ti/modules/titanium/android/AndroidModule.java
	android/modules/android/src/java/ti/modules/titanium/android/notificationmanager/NotificationManagerModule.java
	android/modules/android/src/java/ti/modules/titanium/android/notificationmanager/NotificationProxy.java
	android/modules/app/project.properties
	android/modules/calendar/project.properties
	android/modules/contacts/project.properties
	android/modules/database/project.properties
	android/modules/filesystem/project.properties
	android/modules/geolocation/project.properties
	android/modules/gesture/project.properties
	android/modules/locale/project.properties
	android/modules/map/project.properties
	android/modules/media/project.properties
	android/modules/network/project.properties
	android/modules/network/src/java/ti/modules/titanium/network/TiHTTPClient.java
	android/modules/platform/project.properties
	android/modules/ui/project.properties
	android/modules/ui/src/java/ti/modules/titanium/ui/ScrollViewProxy.java
	android/modules/ui/src/java/ti/modules/titanium/ui/WindowProxy.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiImageView.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIButton.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIImageView.java
	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUITableView.java
	android/modules/utils/project.properties
	android/modules/xml/project.properties
	android/runtime/common/src/java/org/appcelerator/kroll/util/KrollAssetHelper.java
	android/templates/build/theme.xml
	android/titanium/project.properties
	android/titanium/src/java/org/appcelerator/kroll/KrollProxy.java
	android/titanium/src/java/org/appcelerator/titanium/TiBaseActivity.java
	android/titanium/src/java/org/appcelerator/titanium/TiC.java
	android/titanium/src/java/org/appcelerator/titanium/proxy/MenuItemProxy.java
	android/titanium/src/java/org/appcelerator/titanium/proxy/TiWindowProxy.java
	android/titanium/src/java/org/appcelerator/titanium/util/TiFileHelper.java
	android/titanium/src/java/org/appcelerator/titanium/util/TiResponseCache.java
	android/titanium/src/java/org/appcelerator/titanium/view/TiDrawableReference.java
	build/titanium_version.py
	iphone/Classes/APIModule.m
	iphone/Classes/APSHTTPClient/libAPSHTTPClient.a
	iphone/Classes/AudioStreamer/AudioStreamerCUR.m
	iphone/Classes/CalendarModule.m
	iphone/Classes/ContactsModule.m
	iphone/Classes/FilesystemModule.m
	iphone/Classes/ImageLoader.m
	iphone/Classes/KrollObject.m
	iphone/Classes/MGSplitView/MGSplitViewController.m
	iphone/Classes/MediaModule.m
	iphone/Classes/SCListener.m
	iphone/Classes/StreamModule.m
	iphone/Classes/TiAnimation.m
	iphone/Classes/TiApp.m
	iphone/Classes/TiAppPropertiesProxy.m
	iphone/Classes/TiBase.h
	iphone/Classes/TiCalendarCalendar.m
	iphone/Classes/TiFile.m
	iphone/Classes/TiGradient.m
	iphone/Classes/TiMediaVideoPlayerProxy.h
	iphone/Classes/TiNetworkTCPSocketProxy.m
	iphone/Classes/TiProxy.h
	iphone/Classes/TiProxy.m
	iphone/Classes/TiUIAlertDialogProxy.m
	iphone/Classes/TiUIButtonBar.m
	iphone/Classes/TiUIButtonProxy.m
	iphone/Classes/TiUIDashboardViewProxy.m
	iphone/Classes/TiUIImageView.m
	iphone/Classes/TiUILabel.h
	iphone/Classes/TiUILabel.m
	iphone/Classes/TiUIListItem.m
	iphone/Classes/TiUIListSectionProxy.h
	iphone/Classes/TiUIListSectionProxy.m
	iphone/Classes/TiUIListView.h
	iphone/Classes/TiUIListView.m
	iphone/Classes/TiUIListViewProxy.m
	iphone/Classes/TiUIOptionDialogProxy.m
	iphone/Classes/TiUIProgressBar.h
	iphone/Classes/TiUIScrollView.h
	iphone/Classes/TiUIScrollViewProxy.m
	iphone/Classes/TiUIScrollableView.h
	iphone/Classes/TiUIScrollableView.m
	iphone/Classes/TiUIScrollableViewProxy.h
	iphone/Classes/TiUIScrollableViewProxy.m
	iphone/Classes/TiUITableView.h
	iphone/Classes/TiUITableView.m
	iphone/Classes/TiUITableViewProxy.m
	iphone/Classes/TiUITableViewRowProxy.h
	iphone/Classes/TiUITableViewRowProxy.m
	iphone/Classes/TiUITextArea.m
	iphone/Classes/TiUITextField.m
	iphone/Classes/TiUITextWidget.m
	iphone/Classes/TiUIView.m
	iphone/Classes/TiUIWindowProxy.m
	iphone/Classes/TiUIiOSAdView.m
	iphone/Classes/TiUIiOSAdViewProxy.m
	iphone/Classes/TiUIiOSCoverFlowView.m
	iphone/Classes/TiUIiOSProxy.m
	iphone/Classes/TiViewProxy.m
	iphone/Classes/TiWindowProxy.m
	iphone/Classes/UIImage+Resize.m
	iphone/Classes/WebFont.m
	iphone/cli/hooks/package.js
	iphone/iphone/Titanium.xcodeproj/project.pbxproj
	node_modules/node-appc/lib/timodule.js
	package.json

+++ iphone/Classes/TiFilesystemFileProxy.m
@@ -1,6 +1,6 @@
 /**
  * Appcelerator Titanium Mobile
- * Copyright (c) 2009-2010 by Appcelerator, Inc. All Rights Reserved.
+ * Copyright (c) 2009-2014 by Appcelerator, Inc. All Rights Reserved.
  * Licensed under the terms of the Apache Public License
  * Please see the LICENSE included with this distribution for details.
  */
@@ -474,7 +474,7 @@ FILENOOP(setHidden:(id)x);
     u_int8_t value;
     const char* fullPath = [[self path] fileSystemRepresentation];
     
-    int result = getxattr(fullPath, backupAttr, &value, sizeof(value), 0, 0);
+    ssize_t result = getxattr(fullPath, backupAttr, &value, sizeof(value), 0, 0);
     if (result == -1) {
         // Doesn't matter what errno is set to; this means that we're backing up.
         return @YES;
