+++ iphone/Classes/TiUITableViewProxy.m
@@ -158,12 +158,12 @@ USE_VIEW_FOR_CONTENT_SIZE
 -(TiUITableViewRowProxy*)rowForIndex:(NSInteger)index section:(NSInteger*)section
 {
 	int current = 0;
-	int row = index;
+	NSInteger row = index;
 	int sectionIdx = 0;
 	
 	for (TiUITableViewSectionProxy *sectionProxy in sections)
 	{
-		int rowCount = [sectionProxy rowCount];
+		NSInteger rowCount = [sectionProxy rowCount];
 		if (rowCount + current > index)
 		{
 			if (section!=nil)
@@ -188,11 +188,11 @@ USE_VIEW_FOR_CONTENT_SIZE
 	}
 	int section = 0;
 	int current = 0;
-	int row = index;
+	NSInteger row = index;
 	
 	for (TiUITableViewSectionProxy * thisSection in sections)
 	{
-		int rowCount = [thisSection rowCount];
+		NSInteger rowCount = [thisSection rowCount];
 		if (rowCount + current > index)
 		{
 			return [NSIndexPath indexPathForRow:row inSection:section];
@@ -288,7 +288,7 @@ USE_VIEW_FOR_CONTENT_SIZE
 -(TiUITableViewSectionProxy*)sectionForIndex:(NSInteger)index row:(TiUITableViewRowProxy**)rowOut
 {
 	int current = 0;
-	int row = index;
+	NSInteger row = index;
 	int sectionIdx = 0;
 	
 	TiUITableViewRowProxy *rowProxy = nil;
@@ -296,7 +296,7 @@ USE_VIEW_FOR_CONTENT_SIZE
 	
 	for (sectionProxy in sections)
 	{
-		int rowCount = [sectionProxy rowCount];
+		NSInteger rowCount = [sectionProxy rowCount];
 		if (rowCount + current > index)
 		{
 			rowProxy = [sectionProxy rowAtIndex:row];
@@ -453,7 +453,7 @@ USE_VIEW_FOR_CONTENT_SIZE
         
         for (sectionProxy in sections)
         {
-            int rowCount = [sectionProxy rowCount];
+            NSInteger rowCount = [sectionProxy rowCount];
             if (rowCount + current > index)
             {
                 rowProxy = [sectionProxy rowAtIndex:row];
@@ -586,13 +586,12 @@ USE_VIEW_FOR_CONTENT_SIZE
 	
 	TiUITableViewRowProxy *newrow = [self tableRowFromArg:data];
     TiUITableViewActionType actionType = TiUITableViewActionInsertRowBefore;
-//    TiUITableViewSectionProxy *actionSection = section;
     id header = [newrow valueForKey:@"header"];
     if (header != nil) {
         TiUITableViewSectionProxy *newSection = [self sectionWithHeader:header table:table];
         
         // Insert the new section into the array - but, exactly WHERE we insert depends.
-        int sectionIndex = [sections indexOfObject:section];
+        NSInteger sectionIndex = [sections indexOfObject:section];
         if (row.row != 0) {
             sectionIndex++;
         }
@@ -602,7 +601,7 @@ USE_VIEW_FOR_CONTENT_SIZE
 		
         // Thanks to how we track sections, we also need to manually update the index
         // of each section in the array after where the insert will be.
-        for (int i=sectionIndex; i < [sections count]; i++) {
+        for (NSUInteger i=sectionIndex; i < [sections count]; i++) {
             TiUITableViewSectionProxy *updateSection = [sections objectAtIndex:i];
             updateSection.section = updateSection.section + 1;
         }
@@ -615,7 +614,6 @@ USE_VIEW_FOR_CONTENT_SIZE
         
         // Configure the action
         actionType = TiUITableViewActionInsertSectionBefore;
-//        actionSection = newSection;
     }
     else {
 		[section rememberProxy:newrow];	//If we wait until the main thread, it'll be too late!
@@ -666,7 +664,6 @@ USE_VIEW_FOR_CONTENT_SIZE
 	
 	TiUITableViewRowProxy *newrow = [self tableRowFromArg:data];
     TiUITableViewActionType actionType = TiUITableViewActionInsertRowAfter;
-//    TiUITableViewSectionProxy *actionSection = section;
     id header = [newrow valueForKey:@"header"];
     if (header != nil) {
         TiUITableViewSectionProxy *newSection = [self sectionWithHeader:header table:table];
@@ -675,11 +672,11 @@ USE_VIEW_FOR_CONTENT_SIZE
         newSection.section = section.section + 1;
         
         // Insert the new section into the array
-        int sectionIndex = [sections indexOfObject:section] + 1;
+        NSUInteger sectionIndex = [sections indexOfObject:section] + 1;
         
         // Thanks to how we track sections, we also need to manually update the index
         // of each section in the array after where the insert will be.
-        for (int i=sectionIndex; i < [sections count]; i++) {
+        for (NSUInteger i=sectionIndex; i < [sections count]; i++) {
             TiUITableViewSectionProxy *updateSection = [sections objectAtIndex:i];
             updateSection.section = updateSection.section + 1;
         }
@@ -692,7 +689,6 @@ USE_VIEW_FOR_CONTENT_SIZE
         
         // Configure the action
         actionType = TiUITableViewActionInsertSectionAfter;
-//        actionSection = newSection;
     }
     else {
 		[section rememberProxy:newrow];	//If we wait until the main thread, it'll be too late!
@@ -955,7 +951,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
     [(TiUITableView *)[self view] refreshSearchControllerUsingReload:YES];
 }
 
--(int)sectionCount
+-(NSUInteger)sectionCount
 { //TODO: Shouldn't this be in the main thread, too?
 	return [sections count];
 }
@@ -980,7 +976,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 -(void)appendSection:(id)args
 {
 	//Step one: sanity
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	if(argCount < 1){
 		[self throwException:TiExceptionNotEnoughArguments
 				   subreason:@"expected 1 argument, received none" location:CODELOCATION];
@@ -992,7 +988,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 	NSMutableArray * sectionArray = nil;
 
 	if ([appendum isKindOfClass:[NSArray class]]) {
-		int arrayCount = [appendum count];
+		NSUInteger arrayCount = [appendum count];
 		if (arrayCount < 1) {
 			[self throwException:TiExceptionNotEnoughArguments subreason:@"array must not be empty" location:CODELOCATION];
 		}
@@ -1024,7 +1020,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 		NSRange sectionRange = NSMakeRange([sections count], 1);
 		if (sectionArray != nil){
 			sectionRange.length = [sectionArray count];
-			int sectionIndex = sectionRange.location;
+			NSUInteger sectionIndex = sectionRange.location;
 			for (TiUITableViewSectionProxy * thisSection in sectionArray) {
 				[thisSection setSection:sectionIndex++];
 			}
@@ -1047,7 +1043,6 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 				[ourTable insertSections:[NSIndexSet indexSetWithIndexesInRange:sectionRange] withRowAnimation:ourAnimation];
 			} else { //UITableView doesn't know we had 0 sections.
 				[ourTable beginUpdates];
-//				NSRange insertedSectionRange = NSMakeRange(1, sectionRange.length - 1);
 				[ourTable deleteSections:[NSIndexSet indexSetWithIndex:0] withRowAnimation:ourAnimation];
 				[ourTable insertSections:[NSIndexSet indexSetWithIndexesInRange:sectionRange] withRowAnimation:ourAnimation];
 				[ourTable endUpdates];
@@ -1059,7 +1054,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 -(void)deleteSection:(id)args
 {
 	//Step one: sanity
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	if(argCount < 1){
 		[self throwException:TiExceptionNotEnoughArguments
 				   subreason:@"expected 1 argument, received none" location:CODELOCATION];
@@ -1098,8 +1093,8 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 {
 	[self rememberSection:section];
 	TiThreadPerformOnMainThread(^{
-		int oldSectionCount = [sections count];
-		int boundSectionIndex = MIN(sectionIndex, oldSectionCount);
+		NSUInteger oldSectionCount = [sections count];
+		NSUInteger boundSectionIndex = MIN(sectionIndex, oldSectionCount);
 		[section setSection:boundSectionIndex];
 		[sections insertObject:section atIndex:boundSectionIndex];
 
@@ -1120,7 +1115,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 
 -(void)insertSectionAfter:(id)args
 {
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	ENSURE_ARG_COUNT(args, 2)
 
 	NSDictionary * options = nil;
@@ -1146,7 +1141,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 
 -(void)insertSectionBefore:(id)args
 {
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	ENSURE_ARG_COUNT(args, 2)
 	
 	NSDictionary * options = nil;
@@ -1172,7 +1167,7 @@ DEFINE_DEF_PROP(scrollsToTop,@YES);
 
 -(void)updateSection:(id)args
 {
-	int argCount = [args count];
+	NSUInteger argCount = [args count];
 	ENSURE_ARG_COUNT(args, 2)
 	
 	NSDictionary * options = nil;
